Completed the classesof the project

+++ src/main/java/com/anusha/leave_management_system/repository/LeaveBalanceRepository.java
@@ -5,7 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LeaveBalanceRepository extends JpaRepository<LeaveBalance, Integer> {
+public interface LeaveBalanceRepository extends JpaRepository<LeaveBalance, Long> {
     Optional<LeaveBalance> findByEmployeeId(Long employeeId);
 
 }
+
+
+
